Fixes the tests that broke after creation of the UserService service
